Add support for micro patterns, called "figures"
A figure can be defined like this in the yaml file

figures:
    - {name: f1, key: fl, velocity:  65, pattern: .xx}

It can then be referenced from the lanes in a part as +f1 like any
normal note key. The pattern will be played such that it fits exactly in
the duration of one major event.

Add testfiles/beat8.yml for a demonstration of this feature.