Revert "first consume a tick, then play a chord"
This reverts commit 4300371ab1932ca7dc1ecb3e185b27ba98802e9d.

It turned out this commit was wrong in the case to subsequent
parts had different step values. Revert, it makes much more
sense like that.

Rather, the bug that sometimes the first couple of events
are squeezed together seems to correlate with how long the
program needs to start. Should be fixable.